refactor(perl-lexer): split checkpoint module and add edge-case tests (#0000)
